+++ components/form/PublishToggle.tsx
@@ -1,0 +1,20 @@
+"use client";
+
+import { useFormContext } from "react-hook-form";
+
+export function PublishToggle() {
+  const { register } = useFormContext();
+
+  return (
+    <label className="inline-flex items-center cursor-pointer">
+      <input
+        type="checkbox"
+        value=""
+        className="sr-only peer"
+        {...register("status")}
+      />
+      <div className="relative w-11 h-6 bg-gray-200 peer-focus:outline-none peer-focus:ring-4 peer-focus:ring-blue-300 rounded-full peer peer-checked:after:translate-x-full rtl:peer-checked:after:-translate-x-full peer-checked:after:border-white after:content-[''] after:absolute after:top-[2px] after:start-[2px] after:bg-white after:border-gray-300 after:border after:rounded-full after:h-5 after:w-5 after:transition-all peer-checked:bg-blue-600"></div>
+      <span className="ms-3 text-sm font-medium">Publish this post</span>
+    </label>
+  );
+}
